\name{BCEL-package}
\alias{BCEL-package}
\alias{BCEL}
\docType{package}
\title{
\packageTitle{BCEL}
}
\description{
\packageDescription{BCEL}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BCEL}
\packageIndices{BCEL}
Main functions:
exclusivelasso_project(): proximal operator of exclusive lasso.
exclusivelasso(): regression with exclusive lasso penalty.
bcel(): biclustering with exclusive lasso penalty, written with Rcpp.
bcel_stable(): stable method to select models.

}
\author{
\packageAuthor{BCEL}

Maintainer: \packageMaintainer{BCEL}
}
\references{
}
\keyword{ package }
\seealso{
}
\examples{
}
