\name{exclusivelasso}
\alias{exclusivelasso_project}
\alias{exclusivelasso}
\title{Regression with the Exclusive Lasso  Penalty}
\description{
Regression with the Exclusive Lasso Penalty}

\usage{
exclusivelasso_project(z, lambda)                              
exclusivelasso(y, x, group, lambda, iter_max, tol)}

\arguments{
  \item{z}{vector to project.}
  \item{y}{response vector.}
  \item{x}{design matrix.}
  \item{group}{group of variables}
  \item{lambda}{penlaty parameter.}
  \item{iter_max}{max times of iteration.}
  \item{tol}{tolerance for iteration.}
}
\value{
  \code{beta} estimated coefficient.
}
\details{

}
\references{
}
\examples{

}
\author{Yan Zhong}

